move stats cli test to cli folder
